Change base class constructor interface
Description:
The base class constructor interface used ints instead of size_t for
nrows and ncols.
Changed to size_t as this is more appropriate for the argument.

+++ spmv/include/SparseMatrix.hpp
@@ -29,7 +29,7 @@ namespace SpMV
         virtual void _disassembleStorage() =0;
 
     public:
-        SparseMatrix(const int nrows, const int ncols, const MatrixFormat);
+        SparseMatrix(const size_t nrows, const size_t ncols, const MatrixFormat);
         virtual ~SparseMatrix();
 
         inline size_t       getNumRows()     const { return _nrows; };
