+++ spmv/src/SparseMatrix_CSR.cpp
@@ -5,7 +5,7 @@
 namespace SpMV
 {
     template <class fp_type>
-    SparseMatrix_CSR<fp_type>::SparseMatrix_CSR(const int nrows, const int ncols) :
+    SparseMatrix_CSR<fp_type>::SparseMatrix_CSR(const size_t nrows, const size_t ncols) :
          SparseMatrix<fp_type>::SparseMatrix(nrows, ncols, CSR)
     {
         std::cout << "Hello from SparseMatrix_CSR Constructor!" << std::endl;
